requirements.txt
pycryptodome
cryptography
twofish
cffi
python-camellia
rsa